Stage 2 completed

+++ controller/DatabaseOperations.java
@@ -13,9 +13,11 @@ import model.User;
 
 public class DatabaseOperations {
     private DatabaseManager databaseManager = DatabaseManager.getInstance();
-
+    
     private static final DatabaseOperations databaseOperationsInstance = new DatabaseOperations();
-    private Connection connection = databaseManager.getConnection();;
+    private Connection connection = databaseManager.getConnection();
+    
+    private static long latestUserId = 0;
 
     // Private constructor to prevent external instantiation
     private DatabaseOperations() {
@@ -24,6 +26,22 @@ public class DatabaseOperations {
 
     public static DatabaseOperations getInstance() {
         return databaseOperationsInstance;
+    }
+    
+    public void getLatestUserId() {
+    	String sqlQuery = "SELECT MAX(userId) AS maxUserId FROM user";
+    	 try {
+             PreparedStatement preparedStatement = this.connection.prepareStatement(sqlQuery);
+             ResultSet resultSet = preparedStatement.executeQuery();
+
+             if (resultSet.next()) {
+            	 DatabaseOperations.latestUserId = resultSet.getLong("maxUserId");
+             }
+         } catch (Exception e) {
+             e.printStackTrace();
+         }
+
+    	 DatabaseOperations.latestUserId = 0;
     }
 
     public boolean checkUserNameExists(String userName) {
@@ -50,16 +68,18 @@ public class DatabaseOperations {
 
     public boolean saveUserInDatabase(User user) {
     	boolean insertSuccess = false;
-
-        String sqlQuery = "INSERT INTO user VALUES(?,?,?,?)";
+    	DatabaseOperations.latestUserId++;
+        String sqlQuery = "INSERT INTO user VALUES(?,?,?,?,?,?)";
 
         try {
             PreparedStatement preparedStatement = this.connection.prepareStatement(sqlQuery);
             
-            preparedStatement.setString(1, user.getFirstName());
-            preparedStatement.setString(2, user.getLastName());
-            preparedStatement.setString(3, user.getUserName());
-            preparedStatement.setString(4, user.getPassword());
+            preparedStatement.setLong(1, DatabaseOperations.latestUserId);
+            preparedStatement.setString(2, user.getFirstName());
+            preparedStatement.setString(3, user.getLastName());
+            preparedStatement.setString(4, user.getUserName());
+            preparedStatement.setString(5, user.getPassword());
+            preparedStatement.setBoolean(6, user.isVipMember());
             
             int rowsInserted = preparedStatement.executeUpdate();
 
@@ -72,6 +92,32 @@ public class DatabaseOperations {
 
         return insertSuccess;
     }
+    
+    public boolean updateUserInDatabase(User user) {
+    	boolean updateSuccess = false;
+        String sqlQuery = "UPDATE user SET firstName = ?, lastName = ?, userName = ?, password = ?, isVipMember = ? WHERE userId = ?;";
+
+        try {
+            PreparedStatement preparedStatement = this.connection.prepareStatement(sqlQuery);
+            
+            preparedStatement.setString(1, user.getFirstName());
+            preparedStatement.setString(2, user.getLastName());
+            preparedStatement.setString(3, user.getUserName());
+            preparedStatement.setString(4, user.getPassword());
+            preparedStatement.setBoolean(5, user.isVipMember());
+            preparedStatement.setLong(6, user.getUserId());
+            
+            int rowsUpdated = preparedStatement.executeUpdate();
+
+            if (rowsUpdated > 0) {
+            	updateSuccess = true;
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        return updateSuccess;
+    }
 
     public HashMap<String, Object> loginUser(String userName, String password) {
         Connection connection = databaseManager.getConnection();
@@ -79,7 +125,7 @@ public class DatabaseOperations {
         boolean loginSuccess = false;
         User user = null;
 
-        String sqlQuery = "SELECT * FROM User WHERE username = ? AND password = ?";
+        String sqlQuery = "SELECT * FROM user WHERE username = ? AND password = ?";
 
         try {
             PreparedStatement preparedStatement = connection.prepareStatement(sqlQuery);
@@ -91,6 +137,8 @@ public class DatabaseOperations {
             if (resultSet.next()) {
                 loginSuccess = true;
                 user = new User(resultSet.getString("firstName"), resultSet.getString("lastName"), resultSet.getString("userName"), resultSet.getString("password"));
+                user.setUserId(resultSet.getLong("userId"));
+                user.setVipMember(resultSet.getBoolean("isVipMember"));
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -106,7 +154,7 @@ public class DatabaseOperations {
     public boolean addPost(SocialMediaPost post) {
     	boolean insertSuccess = false;
 
-        String sqlQuery = "INSERT INTO socialmediapost VALUES(?,?,?,?,?,?)";
+        String sqlQuery = "INSERT INTO socialmediapost VALUES(?,?,?,?,?,?,?)";
 
         try {
             PreparedStatement preparedStatement = this.connection.prepareStatement(sqlQuery);
@@ -114,9 +162,10 @@ public class DatabaseOperations {
             preparedStatement.setString(1, post.getPostId());
             preparedStatement.setString(2, post.getContent());
             preparedStatement.setString(3, post.getAuthor());
-            preparedStatement.setString(4, post.getLikes());
-            preparedStatement.setString(5, post.getShares());
+            preparedStatement.setLong(4, post.getLikes());
+            preparedStatement.setLong(5, post.getShares());
             preparedStatement.setString(6, post.getDateTime());
+            preparedStatement.setLong(7, post.getUserId());
             
             int rowsInserted = preparedStatement.executeUpdate();
 
@@ -130,15 +179,16 @@ public class DatabaseOperations {
         return insertSuccess;
     }
     
-    public boolean removePost(String postId) {
+    public boolean removePost(String postId, long userId) {
     	boolean deleteSuccess = false;
 
-        String sqlQuery = "DELETE FROM socialmediapost WHERE postId = ?";
+        String sqlQuery = "DELETE FROM socialmediapost WHERE postId = ? AND userId = ?";
 
         try {
             PreparedStatement preparedStatement = this.connection.prepareStatement(sqlQuery);
             
             preparedStatement.setString(1, postId);
+            preparedStatement.setLong(2, userId);
             
             int rowsDeleted  = preparedStatement.executeUpdate();
 
@@ -152,19 +202,20 @@ public class DatabaseOperations {
         return deleteSuccess;
     }
     
-    public SocialMediaPost retrievePost(String postId) {
+    public SocialMediaPost retrievePost(String postId, long userId) {
         
-        String sqlQuery = "SELECT * FROM socialmediapost WHERE postId = ?";
+        String sqlQuery = "SELECT * FROM socialmediapost WHERE postId = ? AND userId = ?";
 
         try {
             PreparedStatement preparedStatement = this.connection.prepareStatement(sqlQuery);
             
             preparedStatement.setString(1, postId);
+            preparedStatement.setLong(2, userId);
             
             ResultSet resultSet = preparedStatement.executeQuery();
 
             if (resultSet.next()) {
-                return new SocialMediaPost(resultSet.getString("postId"), resultSet.getString("content"), resultSet.getString("author"), resultSet.getString("likes"), resultSet.getString("shares"), resultSet.getString("dateTime"));
+                return new SocialMediaPost(resultSet.getString("postId"), resultSet.getString("content"), resultSet.getString("author"), resultSet.getLong("likes"), resultSet.getLong("shares"), resultSet.getString("dateTime"), userId);
             }
             else {
             	return null;
@@ -175,14 +226,15 @@ public class DatabaseOperations {
         return null;
     }
     
-    public ArrayList<SocialMediaPost> getNLikesPost(int numberOfPosts){
-    	String sqlQuery = "SELECT * FROM socialmediapost ORDER BY likes DESC LIMIT ?";
+    public ArrayList<SocialMediaPost> getNLikesPost(int numberOfPosts, long userId){
+    	String sqlQuery = "SELECT * FROM socialmediapost WHERE userId = ? ORDER BY likes DESC LIMIT ?";
     	ArrayList<SocialMediaPost> postList = new ArrayList<>();
     	
         try {
             PreparedStatement preparedStatement = this.connection.prepareStatement(sqlQuery);
             
-            preparedStatement.setInt(1, numberOfPosts);
+            preparedStatement.setLong(1, userId);
+            preparedStatement.setInt(2, numberOfPosts);
             
             ResultSet resultSet = preparedStatement.executeQuery();
             
@@ -191,9 +243,10 @@ public class DatabaseOperations {
             post.setPostId(resultSet.getString("postId"));
             post.setContent(resultSet.getString("content"));
             post.setAuthor(resultSet.getString("author"));
-            post.setLikes(resultSet.getString("likes"));
-            post.setShares(resultSet.getString("shares"));
+            post.setLikes(resultSet.getLong("likes"));
+            post.setShares(resultSet.getLong("shares"));
             post.setDateTime(resultSet.getString("dateTime"));
+            post.setUserId(resultSet.getLong("userId"));
 
             postList.add(post);
 
@@ -205,15 +258,16 @@ public class DatabaseOperations {
     	
     }
     
-    public boolean checkPostIdExists(String postId) {
+    public boolean checkPostIdExists(String postId, long userId) {
         
         boolean postIdExists = false;
-        String sqlQuery = "SELECT * FROM socialmediapost WHERE postId = ?";
+        String sqlQuery = "SELECT * FROM socialmediapost WHERE postId = ? AND userId = ?";
 
         try {
             PreparedStatement preparedStatement = this.connection.prepareStatement(sqlQuery);
             
             preparedStatement.setString(1, postId);
+            preparedStatement.setLong(2, userId);
             
             ResultSet resultSet = preparedStatement.executeQuery();
 
@@ -226,5 +280,60 @@ public class DatabaseOperations {
 
         return postIdExists;
     }
+    
+    public int getCountPostsInRange(int minimumShare, int maximumShare, long userId) {
+    	String sqlQuery = "";
+    	PreparedStatement preparedStatement = null;
+    	try {
+	    	if(maximumShare == -1) {
+	    		sqlQuery = "SELECT COUNT(*) AS postCount FROM socialmediapost WHERE shares >= ? AND userId = ?";
+	    		preparedStatement = this.connection.prepareStatement(sqlQuery);
+	    		
+	    		preparedStatement.setInt(1, minimumShare);
+	            preparedStatement.setLong(2, userId);
+	    	}
+	    	else {
+	    		sqlQuery = "SELECT COUNT(*) AS postCount FROM socialmediapost WHERE shares >= ? AND shares <= ? AND userId = ?";
+	    		preparedStatement = this.connection.prepareStatement(sqlQuery);
+	    		
+	    		preparedStatement.setInt(1, minimumShare);
+	    		preparedStatement.setInt(2, maximumShare);
+	            preparedStatement.setLong(3, userId);
+	    	}
+            
+            ResultSet resultSet = preparedStatement.executeQuery();
+
+            if (resultSet.next()) {
+            	return resultSet.getInt("postCount");
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+    	return 0;
+    }
+    
+    public boolean addVipMember(Long userId) {
+    	boolean updateSuccess = false;
+        String sqlQuery = "UPDATE user SET isVipMember = ? WHERE userId = ?;";
+
+        try {
+            PreparedStatement preparedStatement = this.connection.prepareStatement(sqlQuery);
+            
+            preparedStatement.setBoolean(1, true);
+            preparedStatement.setLong(2, userId);
+            
+            int rowsUpdated = preparedStatement.executeUpdate();
+
+            if (rowsUpdated > 0) {
+            	updateSuccess = true;
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        return updateSuccess;
+    }
+
 }
